// src/app/api/model/productEdit.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductEdit { 
    id?: number;
    name?: string;
    description?: string;
    imageUrl?: string;
    category?: string;
    stock?: number;
    price?: number;
}