// src/app/api/model/billEdit.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BillEdit { 
    id?: number;
    destinatarioDenomSocial?: string;
    destinatarioNif?: string;
    destinatarioDIR?: string;
    destinatarioDireccionFiscal?: string;
    emisorDenomSocial?: string;
    emisorNif?: string;
    emisorDireccionFiscal?: string;
    serie?: string;
    numeroCorrelativo?: string;
    fechaEmision?: string;
    fechaVencimiento?: string;
    concepto?: string;
    albaranes?: string;
    tipoIVA?: string;
    tipoRetencion?: string;
    baseFactura?: string;
    importeFinal?: string;
    formaPago?: string;
    fechaPago?: string;
}