// src/app/api/model/clientEdit.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ClientEdit { 
    id?: number;
    dni?: string;
    nombre?: string;
    empresa?: string;
    pais?: string;
    direccion?: string;
    email?: string;
    telefono?: string;
}